services/requirements.txt
flask
flask_httpauth
Pillow
scikit-image
opencv-python
python-dotenv
requests
